Separated asm files, twiked Makefile and removed warnings.

+++ csrc/idt/idt.c
@@ -27,13 +27,13 @@ void init_idt()
 {
   /* Sets the special IDT pointer up, just like in 'gdt.c' */
   idtp.limit = (sizeof (struct idt_entry) * 256) - 1;
-  idtp.base = &idt;
+  idtp.base = (unsigned int) &idt;
 
   /* Clear out the entire IDT, initializing it to zeros */
   memset(&idt, 0, sizeof(struct idt_entry) * 256);
 
   /* Add any new ISRs to the IDT here using idt_set_gate */
-  
+
   //idt_set_gate();
 
   /* Points the processor's internal register to the new IDT */
